### 2021-11-29 #61 @LukasWallrich
[Re] Groups of diverse problem-solvers outperform groups of highest-ability problem-solvers - most of the time
Thanks for your encouragement in ReScience/call-for-replication#6 - as always, this took longer than expected, but I now managed to complete the replications. I'm very much looking forward to hearing your thoughts.

**Original article:** - two articles that built on each other:
Hong, L., & Page, S. E. (2004). Groups of diverse problem solvers can outperform groups of high-ability problem solvers. Proceedings of the National Academy of Sciences, 101(46), 16385–16389. https://doi.org/10.1073/pnas.0403723101
Grim, P., Singer, D. J., Bramson, A., Holman, B., McGeehan, S., & Berger, W. J. (2019). Diversity, Ability, and Expertise in Epistemic Communities. Philosophy of Science, 86(1), 98–123. https://doi.org/10.1086/701070

**Scientific domain:** Social Psychology (could be called Cognitive Modelling?)
**Suggested editor:** @oliviaguest? (but this does not require specialist knowledge)
