### 2022-03-29 #61 @marieke-woensdregt
Dear @LukasWallrich and @oliviaguest, I am so sorry for the radio silence. I was a bit overwhelmed with other work these past months, but will definitely get to this by the end of this week. My sincere apologies for the delay.

I also thought of some other potential reviewers:
1.  Matt Spike, University of Edinburgh:
https://sites.google.com/site/matspike/
https://github.com/matspike
2. James Winters:
https://j-winters.github.io/
https://github.com/j-winters
3. Fausto Carcassi:
https://thelogicalgrammar.github.io/
https://github.com/thelogicalgrammar
4. Bill Thompson
https://billdthompson.github.io/
https://github.com/billdthompson

These are all people who mostly work on language and cultural evolution, because that happens to be my field, but they are also definitely interested in population-level processes more generally, and all do agent-based modelling and simulations using Python. Hope that helps!
