### 2022-06-01 #61 @LukasWallrich
@oliviaguest I finally got a few names for potential reviewers. I don't know any of them - they were suggested by Francesco Rigoli who convenes a network on computational political psychology. Would you want to invite them or should I reach out? For peer review, I thought it make sense if that comes from you but happy to follow your guidance (incidentally, there might also be suitable reviewers here for #64?)

•	dimitri ognibene: [dimitri.ognibene@unimib.it](mailto:dimitri.ognibene@unimib.it)
•	sven banisch: [sven.banisch@UniVerseCity.de](mailto:sven.banisch@UniVerseCity.de)
•	michael moutoussis: @mmoutou here
•	geert-jan will: [g.j.will@fsw.leidenuniv.nl](mailto:g.j.will@fsw.leidenuniv.nl)
•	david young: @davidjyoung here
•	lion schulz: [lion.schulz@tue.mpg.de](mailto:lion.schulz@tue.mpg.de)
