### 2022-06-01 #61 @mmoutou
Hi people / Dr. Wallrich,

Not sure who is sending this, and Olivia owes me one 😛 for making silly
jokes about a paper of mine not a million miles in spirit from this, but
if you want to get to know me and/or if you want me to help out w.
reviewing, it would be awesome. BTW I am not sure what 'diverse' means
here, but I have a colleague who is ace re diversity, and she's Laura
Dilley ***@***.***> . I am almost sure that she would not send me
to Coventry for mentioning her here.

Bestest
MM
