### 2022-06-01 #61 @marieke-woensdregt
Hi @LukasWallrich  and @oliviaguest,

I am close to finishing my review of this replication. I am just waiting to see if the simulation results come out the same if I re-run the simulations myself. I should be able to send you my review soon!

By the way, how shall I share my review @oliviaguest? Should it be open, e.g. by attaching it as a file to a message in this GitHub thread?
