### 2022-06-02 #61 @oliviaguest
> By the way, how shall I share my review @oliviaguest? Should it be open, e.g. by attaching it as a file to a message in this GitHub thread?

Plain text in this thread would be great. See examples here when you click on "review" http://rescience.github.io/read/ — hope that helps. 😊
