### 2022-06-02 #61 @mmoutou
Yeah that was quite unclear - I didn't realise that even if I didn't log in to github, the contents of my email would be public. There really should be info about this visibility *very prominent* in the email copy that's sent. However I'm still confused - I presume all this means that  @oliviaguest will let me know if I am invited as a reviewer. Presumably the review would include Comments to the Editor and Comments to the Authors, etc., so it wouldn't just be posted here???
