### 2022-06-02 #61 @oliviaguest
@mmoutou if you want to you may delete/edit the email/comment above. It would be wonderful if you wanted to review. Do you have the time and feel up for it?

To be clear, this is not a traditional journal. The review completely takes place here, on GitHub, in public. You may look at examples and instructions here: http://rescience.github.io/ Let me know if you have further questions, @mmoutou! 😌
