### 2022-06-30 #61 @marieke-woensdregt
With my sincere apologies for the endless delay (I kept running into minor issues with my reproduction and then not finding the time to resolve them), below finally my complete review. I've structured it according to the 4 questions to reviewers that are specified on the Rescience C website (points 1-4 below). I'm attaching a zip folder containing all the relevant files I used to reproduce the results.

1.	*The actual replication of the research. The reviewer must evaluate the authors’ claims about a successful replication, applying the standards of the field.*

Overall, the results presented in this article show a clear replication of the original results by Hong & Page (2004) and Grim et al. (2019). Where there are differences in the results between this replication and the original articles, the author addresses these differences adequately with detailed further analysis, and provides convincing explanations.

a.	This is very minor, and I don’t think it would make a difference, but Grim et al. (2019) actually seem to use populations of 9 agents, rather than 10?

b.	In Section 3.1, where it says: “For instance, the standard deviation of the performance of groups of 10 highest‐ability agents with l = 12 observed here was 0.56 while Hong and Page report 0.020.”, I think this should be 1.26 instead of 0.56.

c.	Regarding the difference in diversity values for the high-ability agents in this replication compared to the original Hong & Page article: I wonder whether the author has attempted to contact Hong & Page to ask about this? I’m asking because diversity is the notion that all this work revolves around, and because I read in the guidelines of Rescience C that contacting the original authors when differences are found is encouraged. That said, I do realise that the original article was published nearly 20 years ago, so I’m not sure how likely it is that Hong & Page will be able to dig up the original code (assuming they are willing). Furthermore, the author does provide some convincing evidence for this difference in diversity values *not* being indicative of an implementation difference.

d.	In Figure 2, the range of the scale is quite different from Figures 6 and 9 in Grim et al.: ca. -0.5 – 3.5 in the replication, vs. -0.04 – 0.01 in Grim et al. Does the author have an explanation for this? Does it have something to do with the larger range of l values in the replication?

2.	*Reproducibility of the replication. The reviewers must be able to run the proposed implementation on their computer, and obtain the same results as the authors with the limits of the state of the art of the field.*

I reproduced this replication by running the run_simulation.py script from the Hong_and_Page folder, and the run_simulation_sweep.py script from the Grim_et_al folder, on the computer cluster of my institute. In both cases, the results of my reproduction are very similar to those reported in the replication paper (and accompanying Jupyter notebooks). I’m attaching the pickle files containing my reproduced simulation results, as well as the corresponding Jupyter notebooks showing the analysis results for my reproduction. (The files containing the results of my reproduction can be recognized by the MW_ or _MW prefix or suffix.)

a.	I also tried running a simulation using Google Cloud Engine, using the [pyscript2gce](https://github.com/LukasWallrich/pyscript2gce-production) helper that the author refers to in the README.md, but I got stuck at the step of building a Docker image, which led to the following error when I tried to run the simulation:
ERROR: (gcloud.compute.instances.create-with-container) Missing required argument [--container-image]: You must provide container image

b.	For running the code using the [pyscript2gce](https://github.com/LukasWallrich/pyscript2gce-production) helper, it was also a bit unclear to me what code would need to be added on line 26 of main.py, because at the bottom of that same script it seems to already call up the simulation.py script.

c.	It would be useful to add examples of how to pickle (save) the data at the bottom of each run_simulation.py script. I had to add lines like the following myself, at the bottom of each simulation script:
out.to_pickle("MW_HPmodel_results_"+ datetime.now().strftime("%Y-%m-%d-%H-%M-%S")+".pkl")

d.	In the run_simulation_sweep.py script, I changed the line:
“l”: range(4, 10)
to:
“l”: range(4, 31)
Because if I understood it correctly, this was the settting actually used by the author to produce the simulation results for the paper.

e.	In run_simulation_sweep.py, I also changed:
iterations=1
to:
iterations=100
Again because I believe that’s the setting the author actually used to producer the results for the paper.

f.	With my initial simulation results, I ran into a mismatch with the pickle file protocol that was expected by the Jupyter notebooks for analysis. (If I remember correctly, it was: unsupported pickle protocol: 5). If I remember correctly, I solved this issue by changing the version of pandas in the virtual machine I used on my institute’s computer cluster to run these simulations. After changing pandas to version 1.3.5, everything worked smoothly. Maybe the author can add this type of version information somewhere (e.g. in requirements.txt ?)

g.	In the analysis.ipynb file for the Grim_et_al model, I ran into the following error in the code cell that calculates the win_comps table, as well as the final code cell of the document:
AttributeError: 'float' object has no attribute 'round'
I replaced all calls to .round() as a method to round() as a function instead. This solved the issue.

3.	*Clarity of the code and the instructions for running it. Uncommented or obfuscated code is as bad as no code at all.*

Overall, the code is clear and well-structured, and the instructions for running it in the README.md file are also clear (although see my comments on the [pyscript2gce](https://github.com/LukasWallrich/pyscript2gce-production) helper above). I had a first example simulation of the Hong & Page model running on my own laptop within a matter of minutes. Below are some comments that I think will help make the code and README.md file even clearer:

a.	In the README.md file, where it says “The analyses to replicate Hong & Page can feasibly be run on a laptop in a matter of hours” etc., I think the author means simulations rather than analyses (this phrasing was confusing to me, because it looked like it was referring to the jupyter notebooks that contain the analysis scripts, which in fact run in a matter of minutes).

b.	It would be easier to interact with the folders through the terminal if they didn’t have spaces in the folder names (use e.g. underscores instead).

c.	Docstrings are present for all methods, but relatively minimal. Consider adding a more structured specification of the input arguments and what each method returns (or what changes it makes to an object’s attributes). For widely-used standardised docstring formats, see e.g.: https://realpython.com/documenting-python-code/#docstring-formats ).

d.	In general I would appreciate some more comments in the class and method definitions that walk the reader through the steps that are being taken (e.g. in long methods definitions like draw_agents() and max_search() in HPmodel.py).

e.	Given that not everyone will be familiar with the mesa package, perhaps the author could add a few words on this in the README.md. Just to explain what mesa is and how it is (broadly) used in these scripts. E.g., a brief explanation of what the Agent, Model, BaseScheduler, and BatchRunnerMP classes are useful for.

f.	It is not immediately clear what the parameters “iterations” and “max_steps” in the simulation scripts refer to. Is “iterations” the number of landscapes? If so, what is “max_steps”? (I worry the latter might be confused with the parameter l). Would be good to add comments to the code to explain what each of the parameters are.

g.	Perhaps the author could add a comment to explain how the nr_processes parameter of the BatchRunnerMP() class can be used?

h.	Further to my point above, I initially ran the run_simulation_sweep.py script for the Grim et al. model with the parameter setting: "strategy": "both" (which was the default setting given in the script). But then it turned out that Section 2 of the analysis Jupyter notebook from Grim_et_al expects a dataframe with *only* the “relay” strategy results. (Whereas Section 3 indeed expects a dataframe containing both strategies.) The Jupyter notebook got stuck at code cell 5 (the prep_outputs(res) function) when having to deal with a dataframe that contains results from *both* strategies (and therefore four columns, titled: 'relay_random_solution', 'relay_best_solution', 'tournament_random_solution', 'tournament_best_solution', and an extra column titled ‘strategy’, rather than just two columns titled 'random_solution', 'best_solution'). (I then ran a separate simulation with the setting “strategy”: “relay”, to use in Section 2 of the analysis notebook.) In other words, comments to explain what the possible settings of the parameters are, and what settings the analysis scripts can deal with, would be really helpful. (And/or a more general analysis script that can deal with all possible parameter settings.)

4.	*Clarity and completeness of the accompanying article, in which the authors should clearly state why they think they have replicated the paper (same figures, same graphics, same behavior, etc.) and explain any obstacles they have encountered during the replication work. The reviewers should also consider if the article is sufficiently self-contained.*

Overall, the author has done an impressive job at summarizing the two original models and the replications clearly and concisely. The author has also added more detailed analysis here and there (e.g. the SD’s in Figure 1, and the addition of Table 2), which helps gain more insight into the workings of the model. Below are some minor comments that I think will help make the paper more clear and self-contained.

a.	In the first paragraph of the subsection titled “The basic model”, it might help to make a bit more explicit that the ring of 2,000 values represents 2,000 different possible solutions to a given problem, and that the values represent the quality of the solutions. It might also help to give a concrete example, as Hong & Page’s original article did with the gasoline engine designs.

b.	I’m assuming that the last line of that same paragraph: “When they are in a group, they move together, with each agent in turn moving the entire group to the greatest value their heuristic can identify.”  describes the discussion/conversation process that Hong & Page incorporated in their model (which Grim et al. call the “round-the-table relay dynamic”)? However, this is not entirely clear to me; I think it would help to put it in these discussion/conversation terms more explicitly (to help make the link between this paper and the original papers).

c.	In the subsection titled “Grim and colleagues’ extension and qualification”, I think it would help to emphasise a bit more that different problems are represented by different landscapes, and that Grim et al. used the smoothing procedure in order to make different problem landscapes more related/similar to each other (i.e. to increase the “transportability of best performing heuristics.”). I worry it might otherwise be unclear to the reader how relatedness between nearby values in one problem landscape would have anything to do with other problems, and that they might end up misunderstanding what represents a problem in this model (an entire landscape), and what represents a solution (one point in that landscape).

d.	In the paragraph titled “Implementation”, it says that “most results here are based on 500 landscapes.”. Why not be more specific and say that the Hong & Page replication relied on 500 landscapes, while the Grim et al. replication relied on 100 landscapes?

e.	In the caption of Table 1, please specify what the numbers between parentheses are (presumably standard deviations, as in Hong & Page).

f.	The caption of Figure 1 could do with an explanation of what numbers above 0 and numbers below 0 mean (or a legend + colour difference, as in Grim et al.’s Figure 2).

g.	In the caption of Figure 2, it might be helpful to add “(bigger maximum step length means larger heuristics pool)” at the end of the first sentence.

h.	Where it says “The results closely replicate those presented in Figures 5 to 9 in Grim and colleagues.” On page 5, I think the author means Figures 6 and 9.

[MW_Reproduction.zip](https://github.com/ReScience/submissions/files/9020872/MW_Reproduction.zip)
