+++ Hong_and_Page/HPmodel.py
@@ -1,0 +1,235 @@
+# ABM as proposed by Hong & Page (2004)
+
+from statistics import mean
+from collections import Counter
+from itertools import permutations
+from mesa import Agent, Model
+from mesa.time import BaseScheduler
+import copy
+
+class PSAgent(Agent):
+
+    """Agent for Hong-Page problem-solving model.
+
+    This mesa-agent forms part of a team that searches for the optimal solution in the HPProblem model.
+
+    Attributes:
+        heuristic: List of step lengths this agent can consider.
+        team: Team this agent belongs to. Either "random" or "best" in standard HPProblem.
+        best_solution: Best solution found by this agent so far.
+    """
+    
+    def __init__(self, problem: 'HPProblem', id: str, team: str, k: int = None, l: int = None, heuristic: list = None):
+        """Initializes agent, assigns heuristic and team
+        
+        This initializes an agent to solve a given problem, assigns them a heuristic and a team. The heuristic can
+        either be given as a list of step lengths, or by specifying k (number of different step lengths) and l (maximum step length).
+
+        Args:
+            problem (HPProblem): Problem to solve
+            id (str): Agent ID
+            team (str): Name of the team
+            k (int): Number of different step lengths to consider (e.g., 3 per agent)
+            l (int): Maximum step length to consider (e.g., 10)
+            heuristic (list): Alternative to k and l, the specific heuristic can be given
+        """        
+        super().__init__(id, problem)
+
+        if heuristic == None:
+            self.draw_heuristic(k, l)
+        else:
+            self.heuristic = heuristic    
+
+        self.best_solution = 0
+        self.problem = problem
+        self.team = team
+
+    def draw_heuristic(self, problem: 'HPProblem', k: int, l: int):
+        """Draw heuristic as k random integers up to l"""        
+        self.heuristic = problem.random.sample(range(1,l+1), k)
+
+    def step(self):
+        """Search for highest peak accessible with own heuristic (called by mesa)"""        
+        self.focus, self.best_solution = self.problem.max_search(agent = self)
+
+class HPProblem(Model):
+
+    """Hong-Page problem-solving model to assess performance of different teams.
+
+    This mesa-based agent-based model in an implementation of Hong & Page (2004),
+    in which agents are assigned to teams that strive to find the maximum value
+    in a random landscape.
+
+    Attributes:
+        solution: List of numbers representing 'heights' in the landscape.
+        agent_descriptives: Dict with descriptive statistics for agents in each team (i.e. random and best).
+        best_solution: Dict with best solution found by each team so far.
+
+    Methods:
+        max_search: Evaluate a heuristic across all starting points, or have an agent search from their current location.
+        step: Advance model by one step.    
+    """
+
+    def __init__(self, n: int, k: int, l: int, N_agents: int, seed: int = None, agent_class: Agent = PSAgent): 
+        """Initializes problem, assesses heuristics and creates agent teams
+        
+        Initializes Hong & Page model by drawing the solution landscape, assessing the performance of all heuristics, 
+        and creating best and random agent teams.
+
+        Args:
+            n: Length of solution landscape
+            k: Number of steps to include in each heuristic
+            l: Maximum step size to be considered when drawing heuristics
+            N_agents: Number of agents in each team
+            seed: Random seed for reproducibility
+        """        
+        #Seed automatically set by mesa if provided
+        self.schedule = BaseScheduler(self)
+        self.agent_descriptives = {}
+        self.n = n
+        self.draw_solution(n)
+        self.optimal_solution = max(self.solution)
+        self.best_solution = {"random": 0, "best": 0}
+        self.current_position = {"random": 0, "best": 0}
+        self.draw_agents(k, l, N_agents, agent_class)
+        self.running = True
+
[…]
