+++ Grim_et_al/run_simulation_simple.py
@@ -10,11 +10,18 @@ with httpimport.remote_repo(["Gmodel"], url):
 fixed_params = {"n": 2000,
                "k": 3,
                "N_agents": 10,
-               "l": 12
+               "l": 12,
+               "strategy": "relay"
                }
 variable_params = {
                    "smoothness": list(range(21))
                    }
+
+# Run simulations with various parameter combinations
+# `nr_processes` represents the number of processes to use - typically the number of cores on your machine
+# `iterations` represents the number of landscapes
+# `max_steps` is the maximum number of steps the model takes. Given the design 
+# of the HPModel.step() function, it should never have to take more than 1.
 
 batch_run = BatchRunnerMP(GProblem,
                         16,
@@ -27,3 +34,10 @@ batch_run = BatchRunnerMP(GProblem,
 batch_run.run_all()
 
 out = batch_run.get_model_vars_dataframe()
+
+#Save results to pickle - unless the script is run by `pyscript2gce`
+#in which case it will be saved to Cloud Storage automatically
+import os
+import datetime
+if not 'AM_I_IN_A_DOCKER_CONTAINER' in os.environ:
+    out.to_pickle("HPmodel_results_"+ datetime.now().strftime("%Y-%m-%d-%H-%M-%S")+".pkl")
