+++ Grim_et_al/Gmodel.py
@@ -1,0 +1,145 @@
+# Used this import of HPModel to avoid duplicating files
+import httpimport
+url = "https://gist.githubusercontent.com/LukasWallrich/05f445821fbae694b37a205dc08b2b4f/raw/"
+
+with httpimport.remote_repo(["HPmodel"], url):
+     from HPmodel import HPProblem, PSAgent
+
+# Alternative: download file into same folder, then run
+# from HPmodel import HPProblem, PSAgent
+
+import pandas as pd
+import numpy as np
+from copy import copy
+
+class GrimAgent(PSAgent):
+    """Agent for Hong-Page problem-solving model as extended by Grim et al.
+
+    This agent forms part of a team that searches for the optimal solution in the landscape presented. The
+    only extention from the PSAgent class is that the agent can now search in a tournament strategy, not just in a relay.
+
+    Attributes:
+        heuristic: List of step lengths this agent can consider.
+        team: Team this agent belongs to. Either "random" or "best" in standard HPProblem.
+        best_solution: Best solution found by this agent so far.
+    """
+
+    def step(self):
+        """Search for highest peak accessible with own heuristic, in tournament or relay mode (called by mesa)"""        
+        if self.problem.strategy == "relay":
+           self.focus, self.best_solution = self.problem.max_search(agent = self)
+        if self.problem.strategy == "tournament":
+           self.focus, self.best_solution = self.problem.max_search(agent = self, update = False)
+
+
+class GProblem(HPProblem):
+
+    """Hong & Page problem-solving model, extended by Grim et al, to assess performance of different teams.
+
+    This mesa-based agent-based model implements the model described by Grim et al. (2019),
+    which extends the original model by Hong & Page (2004) to account for landscapes of varying smoothness 
+    and to contrast the original 'relay' model with a 'tournament' model. 
+
+    Attributes:
+        agent_descriptives: Dict with descriptive statistics for agents in each team (i.e. random and best).
+        best_solution: Dict with best solution found by each team so far.
+
+    Methods:
+        max_search: Evaluate a heuristic across all starting points, or have an agent search from their current location.
+        step: Advance model by one step.    
+    """
+
+    def __init__(self, n: int, k: int, l: int, N_agents: int, smoothness: int, seed:int = None, strategy: str = "relay", agent_class: PSAgent = GrimAgent):
+        """Initializes problem, assesses heuristics and creates agent teams
+        
+        Initializes Hong & Page model as extended by Grim et al by drawing a smoothed solution landscape, 
+        assessing the performance of all heuristics, and creating best and random agent teams.
+
+        Args:
+            n: Length of solution landscape
+            k: Number of steps to include in each heuristic
+            l: Maximum step size to be considered when drawing heuristics
+            N_agents: Number of agents in each team
+            smoothness: Smoothness of solution landscape, i.e. the average distance between randomly selected points between which the solution is interpolated.
+            seed: Random seed for reproducibility
+            strategy: 
+              Problem solving strategy for teams. Can be 'relay', 'tournament' or 'both. In 'relay' mode, agents sequentially search for improvements,
+              if they find one, their entire team moves to their improved solution and the next agent continues from there. In 'tournament' mode, each
+              agent independently searches for improvements, and then the teams move to the best solution found in that round.
+    
+        """        
+        self.draw_G_solution(n, smoothness)
+        super().__init__(n, k, l, N_agents, seed, agent_class = agent_class)
+        self.strategy = strategy
+
+    def draw_solution(self, n):
+        """Overridden in favor of draw_G_solution"""        
+        pass
+
+    def draw_G_solution(self, n: int, smoothness: int) -> None:
[…]
