+++ requirements.txt
@@ -1,5 +1,5 @@
 jupyter
 matplotlib
-pandas
+pandas >= 1.3.5
 mesa
 httpimport

[…]
