### 2022-08-18 #61 @oliviaguest
@LukasWallrich I am sorry this is taking so long (not that usual to be this slow, but it's both summer and pandemic). I emailed with @thelogicalgrammar, but I assume they may be on holiday.

If you have any ideas for who else to invite, I can try them too. But I think for now, we have to wait out August and try people again in September.
