### 2022-08-29 #61 @marieke-woensdregt
Dear @LukasWallrich (and @oliviaguest),

Thank you for the clear and thorough response to my review. Please find below my response to the revisions, only to those points where a response is relevant (to all other points, my response is some variant of "Great! Thanks for clarifying/adding/editing.").

1.a) Great! As far as I'm concerned a footnote is sufficient. I don't have any reason to believe simulations with populations of 9 agents would look different than with 10.

1.c) Excellent! Great to hear that you were able to get clarity on this. Where you wrote _"However, this is simply due to the fact that Hong and Page’s standard deviations are those of the means, i.e. what are more commonly called standard errors, while I present standard deviations of the observed variables."_ , perhaps add that you know this from "personal communication" with the authors? Given that the caption of Table 1 in Hong & Page just says "standard deviations"?

2.a) I think what went wrong when I tried to build the Docker image is reflected in the following error message from the build log (I'm also attaching the entire build log file to this message):
[build_log.txt](https://github.com/ReScience/submissions/files/9448304/build_log.txt)

`denied: Token exchange failed for project 'rescience-review-diversity-abm'. Caller does not have permission 'storage.buckets.create'. To configure permissions, follow instructions at: https://cloud.google.com/container-registry/docs/access-control
ERROR: push attempt 1 detected failure, retrying: step exited with non-zero status: 1`

Perhaps this had something to do with the fact that I copied the relevant scripts over into a github repo of my own (the one called "rescience-review-diversity-abm": https://github.com/marieke-woensdregt/review_diversity_abm ). I don't fully remember now, but I believe I did that because I wanted to make some changes inside the scripts (e.g. change the email address in config.conf to my own, so you wouldn't get update emails saying my simulations had finished running ;) )
But once I got that error message when trying to build the Docker image, I simply decided to switch to running the simulations on my institute's cluster, rather than figuring out the permissions stuff.

2.c) Good solution! But I happened to notice the output filenames now all start with the same "HPmodel_results_" prefix, also in the Grim simulation scripts?

2.f) Yes, good point. Yes, I indeed had to _upgrade_ the version of pandas in my virtual environment compared to the default one of my institute's cluster (which apparently is still at version 0.25.3  :'-) )

3.c) The docstrings look great now!

3.g) Great! Perhaps just consider writing `nr_processes = 16` rather than just `16` in calls to the BatchRunnerMP() method in the simulation scripts, because the comment above now explains what the `nr_processes` parameter is for, but then that parameter/argument name is not actually visible below.

4.b) It makes sense that you don't want to refer to the relay strategy as "communication". I think it's described clearly now in the paper.

4.c) Great, this is clear now!
