### 2022-08-31 #61 @thelogicalgrammar
Hi everyone,

My apologies for taking so long to write this. Since Marieke covered so much ground already and I feel the project is very close to being done, I mostly just want to add some comments about the code.

**1. The actual replication of the research. The reviewer must evaluate the authors’ claims about a successful replication, applying the standards of the field.**

The authors focus on two papers, Hong & Page (2004) and Grim et al (2019). The results of the two papers are successfully replicated, minus some variations from the original results that are explored by the authors of the replication.

**2. Reproducibility of the replication. The reviewers must be able to run the proposed implementation on their computer, and obtain the same results as the authors with the limits of the state of the art of the field.**

I could reproduce the results presented in the paper, with some minor changes to the code:
- Strangely, mesa did not work when installed through conda, but did work when installed via pip. Maybe worth mentioning this somewhere.
- in Hong_and_Page/run_simulation.py "import datetime" should be "from datetime import datetime", or alternatively "datetime.now()" should be "datetime.datetime.now()"

**3. Clarity of the code and the instructions for running it. Uncommented or obfuscated code is as bad as no code at all.**

- I found the folder structures somewhat unclear and the number of files a bit overwhelming at first. It would be great if the README gave a slightly more detailed description of the individual files.
- Many lines of code were very long which forced code folding in a way that gave a chaotic visual appearance to the code. Example:
```
res_random = res.agent_descriptives.apply(pd.Series).random.apply(pd.Series).rename(mapper = partial(renamer, prefix = "random_"), axis = "columns")
```
can be rewritten as:
```
res_random = (
    res
    .agent_descriptives
    .apply(pd.Series)
    .random.apply(pd.Series)
    .rename(
        mapper = partial(renamer, prefix = "random_"),
        axis = "columns"
    )
)
```
- In some cases indentation was not consistent with 2 or 4 spaces (e.g. line 233 and 235 in HPmodel.py, lines 31 and 33 in Gmodel.py)
- Do not use keywords as argument names. E.g. PSAgent has 'id' as an argument.
- I thought the type annotations for the functions were nice!
- "heuristic == None" --> "heuristic is None"
- The documentation for some classes is incomplete. E.g., in HPProblem only methods 'max_search' and 'step' are reported.
- Please split long list comprehensions into multiple lines
- Very minor: Why "import copy" rather than "from copy import copy"?
- run_simulation.py in the top folder is completely uncommented
- HPmodel, line 77: PSAgent is undefined, replace with PS
- Please move all imports at the top of the file!
- The notebooks are generally well-structured, but line-by-line comments on the code are almost entirely missing.
- Please move all functions in the notebooks to the top of the relevant sections and all imports to the first cell
- Please format the definition of "best_teams" in the notebook

**4. Clarity and completeness of the accompanying article, in which the authors should clearly state why they think they have replicated the paper (same figures, same graphics, same behavior, etc.) and explain any obstacles they have encountered during the replication work. The reviewers should also consider if the article is sufficiently self-contained.**

The authors have done an excellent job at summarizing the previous work and presenting their replication, especially for such a short summary. Just a couple points:
- Figure 1 is missing the x and y labels.
- Blue and purple are very similar colors, if it's easy to do so it would be good to have colors with higher contrast for the legend in Figure 1.
- This is very minor, but why is the definition of the figures so low? A dpi of at least 300 is generally a good default. This does not impede reading the figure, so do not feel like you need to fix it if it would take a long time to do so.
