### 2022-09-09 7d0dd27
Expand description of files and warn of conda issue
