+++ Hong_and_Page/HPmodel.py
@@ -229,11 +229,11 @@ class HPProblem(Model):
         for i in range(self.n):
             self.current_position = dict.fromkeys(self.current_position, i)
             while True:
-               old_solution = self.best_solution
-               self.schedule.step()
-               if old_solution == self.best_solution:
-                solutions.append(copy.copy(self.best_solution))
-                break
+                old_solution = self.best_solution
+                self.schedule.step()
+                if old_solution == self.best_solution:
+                    solutions.append(copy.copy(self.best_solution))
+                    break
         self.best_solution = self.dict_mean(solutions)
         self.running = False    
 
