### 2022-09-09 61b8af0
Complete model formatting and docstrings
