+++ Grim_et_al/Gmodel.py
@@ -1,9 +1,10 @@
 # Used this import of HPModel to avoid duplicating files
 import httpimport
+
 url = "https://gist.githubusercontent.com/LukasWallrich/05f445821fbae694b37a205dc08b2b4f/raw/"
 
 with httpimport.remote_repo(["HPmodel"], url):
-     from HPmodel import HPProblem, PSAgent
+    from HPmodel import HPProblem, PSAgent
 
 # Alternative: download file into same folder, then run
 # from HPmodel import HPProblem, PSAgent
@@ -11,6 +12,7 @@ with httpimport.remote_repo(["HPmodel"], url):
 import pandas as pd
 import numpy as np
 from copy import copy
+
 
 class GrimAgent(PSAgent):
     """Agent for Hong-Page problem-solving model as extended by Grim et al.
@@ -25,11 +27,13 @@ class GrimAgent(PSAgent):
     """
 
     def step(self):
-        """Search for highest peak accessible with own heuristic, in tournament or relay mode (called by mesa)"""        
+        """Search for highest peak accessible with own heuristic, in tournament or relay mode (called by mesa)"""
         if self.problem.strategy == "relay":
-            self.focus, self.best_solution = self.problem.max_search(agent = self)
+            self.focus, self.best_solution = self.problem.max_search(agent=self)
         if self.problem.strategy == "tournament":
-            self.focus, self.best_solution = self.problem.max_search(agent = self, update = False)
+            self.focus, self.best_solution = self.problem.max_search(
+                agent=self, update=False
+            )
 
 
 class GProblem(HPProblem):
@@ -37,8 +41,8 @@ class GProblem(HPProblem):
     """Hong & Page problem-solving model, extended by Grim et al, to assess performance of different teams.
 
     This mesa-based agent-based model implements the model described by Grim et al. (2019),
-    which extends the original model by Hong & Page (2004) to account for landscapes of varying smoothness 
-    and to contrast the original 'relay' model with a 'tournament' model. 
+    which extends the original model by Hong & Page (2004) to account for landscapes of varying smoothness
+    and to contrast the original 'relay' model with a 'tournament' model.
 
     Attributes:
         agent_descriptives: Dict with descriptive statistics for agents in each team (i.e. random and best).
@@ -46,13 +50,23 @@ class GProblem(HPProblem):
 
     Methods:
         max_search: Evaluate a heuristic across all starting points, or have an agent search from their current location.
-        step: Advance model by one step.    
+        step: Advance model by one step.
     """
 
-    def __init__(self, n: int, k: int, l: int, N_agents: int, smoothness: int, seed:int = None, strategy: str = "relay", agent_class: PSAgent = GrimAgent):
+    def __init__(
+        self,
+        n: int,
+        k: int,
+        l: int,
+        N_agents: int,
+        smoothness: int,
+        seed: int = None,
+        strategy: str = "relay",
+        agent_class: PSAgent = GrimAgent,
+    ):
         """Initializes problem, assesses heuristics and creates agent teams
-        
-        Initializes Hong & Page model as extended by Grim et al by drawing a smoothed solution landscape, 
+
+        Initializes Hong & Page model as extended by Grim et al by drawing a smoothed solution landscape,
         assessing the performance of all heuristics, and creating best and random agent teams.
 
         Args:
@@ -62,84 +76,111 @@ class GProblem(HPProblem):
             N_agents: Number of agents in each team
             smoothness: Smoothness of solution landscape, i.e. the average distance between randomly selected points between which the solution is interpolated.
             seed: Random seed for reproducibility
-            strategy: 
+            strategy:
               Problem solving strategy for teams. Can be 'relay', 'tournament' or 'both. In 'relay' mode, agents sequentially search for improvements,
[…]
