+++ Hong_and_Page/HPmodel.py
@@ -5,7 +5,8 @@ from collections import Counter
 from itertools import permutations
 from mesa import Agent, Model
 from mesa.time import BaseScheduler
-import copy
+from copy import copy
+
 
 class PSAgent(Agent):
 
@@ -17,11 +18,22 @@ class PSAgent(Agent):
         heuristic: List of step lengths this agent can consider.
         team: Team this agent belongs to. Either "random" or "best" in standard HPProblem.
         best_solution: Best solution found by this agent so far.
+
+    Methods:
+        step: Advance by one step - called by mesa.
     """
-    
-    def __init__(self, problem: 'HPProblem', id: str, team: str, k: int = None, l: int = None, heuristic: list = None):
+
+    def __init__(
+        self,
+        problem: "HPProblem",
+        id: str,
+        team: str,
+        k: int = None,
+        l: int = None,
+        heuristic: list = None,
+    ):
         """Initializes agent, assigns heuristic and team
-        
+
         This initializes an agent to solve a given problem, assigns them a heuristic and a team. The heuristic can
         either be given as a list of step lengths, or by specifying k (number of different step lengths) and l (maximum step length).
 
@@ -32,25 +44,26 @@ class PSAgent(Agent):
             k (int): Number of different step lengths to consider (e.g., 3 per agent)
             l (int): Maximum step length to consider (e.g., 10)
             heuristic (list): Alternative to k and l, the specific heuristic can be given
-        """        
+        """
         super().__init__(id, problem)
 
-        if heuristic == None:
-            self.draw_heuristic(k, l)
+        if heuristic is None:
+            self.__draw_heuristic(k, l)
         else:
-            self.heuristic = heuristic    
+            self.heuristic = heuristic
 
         self.best_solution = 0
         self.problem = problem
         self.team = team
 
-    def draw_heuristic(self, problem: 'HPProblem', k: int, l: int):
-        """Draw heuristic as k random integers up to l"""        
-        self.heuristic = problem.random.sample(range(1,l+1), k)
+    def __draw_heuristic(self, problem: "HPProblem", k: int, l: int):
+        """Draw heuristic as k random integers up to l"""
+        self.heuristic = problem.random.sample(range(1, l + 1), k)
 
     def step(self):
-        """Search for highest peak accessible with own heuristic (called by mesa)"""        
-        self.focus, self.best_solution = self.problem.max_search(agent = self)
+        """Search for highest peak accessible with own heuristic (called by mesa)"""
+        self.focus, self.best_solution = self.problem.max_search(agent=self)
+
 
 class HPProblem(Model):
 
@@ -67,13 +80,25 @@ class HPProblem(Model):
 
     Methods:
         max_search: Evaluate a heuristic across all starting points, or have an agent search from their current location.
-        step: Advance model by one step.    
+        draw_agents: Generate teams of agents (random and best)
+        generate_heuristics: Create heuristics (set of step sizes to be considered)
+        evaluate_heuristics: Calculate average score achieved by a given heuristic
+        assess_hp_diversity: Calculate diversity between two heuristics as defined by Hong & Page
+        step: Advance model by one step.
     """
 
-    def __init__(self, n: int, k: int, l: int, N_agents: int, seed: int = None, agent_class: Agent = PSAgent): 
+    def __init__(
+        self,
+        n: int,
+        k: int,
+        l: int,
+        N_agents: int,
+        seed: int = None,
+        agent_class: Agent = PSAgent,
+    ):
         """Initializes problem, assesses heuristics and creates agent teams
-        
-        Initializes Hong & Page model by drawing the solution landscape, assessing the performance of all heuristics, 
+
[…]
