### 2022-09-09 ef31da7
Make draw_solution public again
