### 2022-09-09 c958dad
Add footnote and increase dpi
