+++ manuscript/content.tex
@@ -21,7 +21,7 @@ The model formulation used in this paper is the same as that used in the origina
 
 \subsubsection{Implementation}
 
-To replicate the two papers, and support future research, I implemented the model in Python, using the mesa framework.\supercite{kazil2020utilizing} This yields readable explicit code, and the derivation of the 'Grim-model' from the 'Hong and Page' model might serve as an example for future model adaptations. While Hong and Page (2004) reported results based on 50 random landscapes and Grim and colleagues relied on 100 runs, the key replications here are based on 500 landscapes. Broader parameter sweeps that establish boundary conditions for the results reported by Grim et al. are based on 100 landscapes for each setting. In the supplementary material, instructions for deploying these scripts to Google Cloud Engine are provided, which allows to reproduce all numerical results in just over 24 hours on 32 cores.
+To replicate the two papers, and support future research, I implemented the model in Python, using the mesa framework\supercite{kazil2020utilizing}. This yields readable explicit code, and the derivation of the 'Grim-model' from the 'Hong and Page' model might serve as an example for future model adaptations. While Hong and Page (2004) reported results based on 50 random landscapes and Grim and colleagues relied on 100 runs, the key replications here are based on 500 landscapes. Broader parameter sweeps that establish boundary conditions for the results reported by Grim et al. are based on 100 landscapes for each setting. In the supplementary material, instructions for deploying these scripts to Google Cloud Engine are provided, which allows to reproduce all numerical results in just over 24 hours on 32 cores.
 
 \subsubsection{Parameters}
 
@@ -56,7 +56,7 @@ To enable direct comparisons between the replication results and the original pa
     \end{table}
 
     
-The replication confirmed the pattern of results observed by Hong and Page, as can be seen in Table \ref{tab:HP}. Random groups of agents outperformed groups of only the high-ability agents in each of the four scenarios, and the observed performances were generally similar here and in the target paper. However, there are two notable divergences. Firstly, the \textit{standard deviations} reported by Hong and Page and those observed here vary by orders of magnitude. For instance, the standard deviation of the performance of groups of 10 highest-ability agents with \textit{l} = 12 observed here was 1.26 while Hong and Page report 0.020. However, this is simply due to the fact that Hong and Page's standard deviations are those of the means, i.e. what are more commonly called standard errors, while I present standard deviations of the observed variables.\footnote{Standard errors are standard deviations divided by the square root of the sample size. Where Hong & Page report the standard error of the diversity of random agents as 0.232, this can thus be converted to a standard deviation by multiplying with the square root of 50, resulting in 1.64. Given the small sample size, this does not differ substantially from the standard deviation of 2.52 observed in my simulations. Their standard errors for the solutions are based on 100,000 runs (50 runs starting from 2000 locations), so that their standard error of 0.007 for random teams indicates a standard deviation of 0.7, again fairly close to my result of 0.56. I am not presenting standard errors here, given that they are a function of the number of runs and would thus not be comparable when running a number of replications in line of what can be expected with current computing resources. Standard deviations, on the other hand, can give a useful indication of the distribution of results arounds the means reported. Additionally, the fact that the runs to obtain performance figures are not independent but clustered within landscapes suggests that the calculation method used by Hong & Page does not give a reliable indication of precision for the performance measures in any case. Clustered standard errors would be needed for that.} \\
[…]
