+++ manuscript/article-metadata.tex
@@ -18,16 +18,16 @@
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{Social Psychology}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
-\def \articleYEAR{2021}
+\def \articleYEAR{2022}
 \def \reviewURL{}
-\def \articleABSTRACT{Problem-solving teams are becoming increasingly diverse, which has been suggested to improve performance. Pioneering computational work by Hong \& Page (2004) suggested that diversity indeed trumps ability when it comes to finding the peaks in a random landscape. Recently, Grim et al. (2019) extended the model to consider more structured landscapes and suggested that the original claim only holds under limited circumstances. In this paper, I replicate the main findings of the two works, and provide modular, extensible code to facilitate future research into this paradigm. Overall, the replications are successful, even though a closer look at the data suggests that the limitations highlighted by Grim et al. are weaker than initially presented, so that diversity on the whole still appears to trump ability within this paradigm. The code for the agent-based models and my analyses is available on https://github.com/LukasWallrich/diversity_abm_replication.}
+\def \articleABSTRACT{Problem-solving teams are becoming increasingly diverse, which has been suggested to improve performance. Pioneering computational work by Hong \& Page (2004) suggested that diversity indeed trumps ability when it comes to finding the peaks in a random landscape. Recently, Grim and colleagues (2019) extended the model to consider more structured landscapes and suggested that the original claim only holds under limited circumstances. In this paper, I replicate the main findings of the two works, and provide modular, extensible code to facilitate future research into this paradigm. Overall, the replications were successful, even though a closer look at the data suggests that the limitations highlighted by Grim and colleagues are less severe than initially presented, so that diversity on the whole still appears to trump ability within this paradigm. The code for the agent-based models and my analyses is available on https://github.com/LukasWallrich/diversity abm replication.}
 \def \replicationCITE{Grim, P., Singer, D. J., Bramson, A., Holman, B., McGeehan, S., & Berger, W. J. (2019). Diversity, ability, and expertise in epistemic communities. Philosophy of Science, 86(1), 98-123.}
 \def \replicationBIB{grim_diversity_2019}
-\def \replicationURL{https://www.researchgate.net/publication/328150961_Diversity_Ability_and_Expertise_in_Epistemic_Communities/link/605b6a0aa6fdccbfea048495/download}
+\def \replicationURL{https://www.researchgate.net/publication/328150961\_Diversity\_Ability\_and\_Expertise\_in\_Epistemic\_Communities/link/605b6a0aa6fdccbfea048495/download}
 \def \replicationDOI{10.1086/701070}
 \def \contactNAME{Lukas Wallrich}
 \def \contactEMAIL{l.wallrich@bbk.ac.uk}
-\def \articleKEYWORDS{agent-based modeling, diversity, Python}
+\def \articleKEYWORDS{agent-based modeling, diversity, problem-solving, Python}
 \def \journalNAME{ReScience C}
 \def \journalVOLUME{4}
 \def \journalISSUE{1}
