### 2022-09-09 #61 @LukasWallrich
@marieke-woensdregt, thank you so much for these further helpful comments. I have implemented all the suggested fixes (for completeness included below). The only aspect where I would appreciate your further thoughts is 2a - there, I am not sure if I have fixed it and if further work is needed at present. My tendency would be to leave `pyscript2gce` as is, given that it is tangential to this replication - but pls see below.

> 1.c) Excellent! Great to hear that you were able to get clarity on this. Where you wrote _"However, this is simply due to the fact that Hong and Page’s standard deviations are those of the means, i.e. what are more commonly called standard errors, while I present standard deviations of the observed variables."_ , perhaps add that you know this from "personal communication" with the authors? Given that the caption of Table 1 in Hong & Page just says "standard deviations"?

Good point. I added that into the footnote.

> 2.a) I think what went wrong when I tried to build the Docker image is reflected in the following error message from the build log (I'm also attaching the entire build log file to this message): [build_log.txt](https://github.com/ReScience/submissions/files/9448304/build_log.txt)
>
> `denied: Token exchange failed for project 'rescience-review-diversity-abm'. Caller does not have permission 'storage.buckets.create'. To configure permissions, follow instructions at: https://cloud.google.com/container-registry/docs/access-control ERROR: push attempt 1 detected failure, retrying: step exited with non-zero status: 1`
>
> Perhaps this had something to do with the fact that I copied the relevant scripts over into a github repo of my own (the one called "rescience-review-diversity-abm": https://github.com/marieke-woensdregt/review_diversity_abm ). I don't fully remember now, but I believe I did that because I wanted to make some changes inside the scripts (e.g. change the email address in config.conf to my own, so you wouldn't get update emails saying my simulations had finished running ;) ) But once I got that error message when trying to build the Docker image, I simply decided to switch to running the simulations on my institute's cluster, rather than figuring out the permissions stuff.

Thanks for sharing this. It seems that there is an issue with the specific permissions of the user - hard to troubleshoot remotely. I have added a command to activate the Cloud Storage API explicitly to the README which might help. @thelogicalgrammar, did you run this using the pyscript2gce helper? *Given that this is entirely tangential to the replication, I am not sure if we need to resolve it?* (Happy to troubleshoot further if someone ever opens an issue in pyscript2gce ...)

>
> 2.c) Good solution! But I happened to notice the output filenames now all start with the same "HPmodel_results_" prefix, also in the Grim simulation scripts?

Fixed, good catch ... too much copy-pasting.

> 2.f) Yes, good point. Yes, I indeed had to _upgrade_ the version of pandas in my virtual environment compared to the default one of my institute's cluster (which apparently is still at version 0.25.3 :'-) )

Perfect - then the minimum version in requirements.txt is appropriate.

> 3.g) Great! Perhaps just consider writing `nr_processes = 16` rather than just `16` in calls to the BatchRunnerMP() method in the simulation scripts, because the comment above now explains what the `nr_processes` parameter is for, but then that parameter/argument name is not actually visible below.

Done.
