### 2022-09-09 #61 @LukasWallrich
Dear @thelogicalgrammar,

Thank you so much for taking the time to review this and to provide such helpful feedback.

Firstly, just one point where I would appreciate clarification.

> * HPmodel, line 77: PSAgent is undefined, replace with PS

I don't quite understand this point - the `PSAgent` class is defined right above.

Apart from that, I implemented all your very helpful suggestions. Specifically:

> * Strangely, mesa did not work when installed through conda, but did work when installed via pip. Maybe worth mentioning this somewhere.

I added a note regarding this to the README.

> * in Hong_and_Page/run_simulation.py "import datetime" should be "from datetime import datetime", or alternatively "datetime.now()" should be "datetime.datetime.now()"

Thanks - changed.

> **3. Clarity of the code and the instructions for running it. Uncommented or obfuscated code is as bad as no code at all.**
>
> * I found the folder structures somewhat unclear and the number of files a bit overwhelming at first. It would be great if the README gave a slightly more detailed description of the individual files.

I have expanded on the description

> * Many lines of code were very long which forced code folding in a way that gave a chaotic visual appearance to the code. Example:
> * In some cases indentation was not consistent with 2 or 4 spaces (e.g. line 233 and 235 in HPmodel.py, lines 31 and 33 in Gmodel.py)
> * Please split long list comprehensions into multiple lines
> * Do not use keywords as argument names. E.g. PSAgent has 'id' as an argument.

Thank you! I'm very new to Python, so wasn't aware of how easy it is to add reasonable line breaks. Now I did that and also used black to format the code overall.

> * "heuristic == None" --> "heuristic is None"

Changed.

> * The documentation for some classes is incomplete. E.g., in HPProblem only methods 'max_search' and 'step' are reported.

Good point. I had only documented the important methods, but to follow good practice, I now changed some methods to internal and documented the others.

> * Very minor: Why "import copy" rather than "from copy import copy"?
> * Please move all imports at the top of the file!

Both changed.

> * run_simulation.py in the top folder is completely uncommented

This file is only meant to be used to run simulations on Google Cloud Engine; I wanted to leave it there to show how that workflow can work. The actual simulation code for the article is in the subfolders. I now highlighted this in the README and added a comment at the top of run_simulation.py.

> * The notebooks are generally well-structured, but line-by-line comments on the code are almost entirely missing.

I now added some code-comments and revised some headings / text so that it should be clear what is happening in each cell.

> * Please move all functions in the notebooks to the top of the relevant sections and all imports to the first cell
> * Please format the definition of "best_teams" in the notebook

Done.
