+++ manuscript/header.tex
@@ -80,9 +80,9 @@
 
 
 %% --- Abstract ---------------------------------------------------------------
-%% \ifdefempty{\articleABSTRACT}{}{
-%%   {\small \sffamily \textbf{Abstract} \articleABSTRACT \par}
-%% }
+\ifdefempty{\articleABSTRACT}{}{
+   {\small \sffamily \textbf{Abstract} \articleABSTRACT \par}
+ }
 
 
 %% --- Replication ------------------------------------------------------------
