### 2022-10-10 #61 @thelogicalgrammar
Dear @LukasWallrich,

Thanks for the detailed answer! I am happy with the edits, everything looks ready to me.

> > * HPmodel, line 77: PSAgent is undefined, replace with PS
>
> I don't quite understand this point - the `PSAgent` class is defined right above.

PSAgent was referred to as "PS" later in the code, but this seems to have been solved already.

Best,
Fausto
