### 2022-10-19 #61 @marieke-woensdregt
Dear @LukasWallrich,

Apologies for the late reply. Thank you for the final fixes. Those all look good!
And regarding my point 2a: I agree with your view that the pyscript2gce helper is tangential to the replication. I think the fact that you added a command to activate the Cloud Storage API explicitly to the README should help. Unfortunately, the free trial of Google Cloud Engine that I used to first try and run this replication myself for my initial review has now run out, so I cannot check directly whether this addition would resolve the issue I had initially. But I agree that _if_ there is still an issue here, it does not need to be resolved for the purposes of this Rescience C paper and replication. I agree that you can just leave it (and of course look into it if someone were to at some point open an issue with the pyscript2gce helper on github, as you say).

In sum, all my comments have been adequately addressed, and I consider this paper and replication ready for publication!

Best wishes,
Marieke
