### 2022-10-20 2ad7bc5
Minor cleanup
