+++ README.md
@@ -28,7 +28,7 @@ Note that there has been an issue with installing mesa from conda - so we recomm
 
 The simulations to replicate Hong & Page can feasibly be run on a laptop in a matter of hours. However, by the time it comes to the parameter sweep needed for the strategy comparisons in the Grim et al. paper, 56,700 runs are needed, which each involve the evaluation of up to 24,360 possible heuristics for 2000 starting locations. On a 32-core Virtual Machine, this took just 22.5 hours. However, it probably took me as long to figure out how to conveniently deploy such scripts to GCE.
 
-When I failed to find a simple solution, I created the [pyscript2gce](https://github.com/LukasWallrich/pyscript2gce-production) helper, which creates a Docker container that executes a script when launched and saves the results. Once set up, all it takes to run a script is to push an update to a specified file (here: `run_simulation.py`) and start up a VM with a single line terminal command. For that, [a GitHub action](https://github.com/LukasWallrich/diversity_abm_replication/blob/main/.github/workflows/push_gist.yml) pushes changes to that file automatically to a Gist, which is then accessed by the VM, based on [this version](https://github.com/LukasWallrich/pyscript2gce-production/releases/tag/Diversity-ABM-replication) of pyscript2gce. The README of pyscript2gce details how this can be set up. Note that do not need to change anything in the Python code in `pyscript2gce` except for the link to the Gist in `run_simulation.py` if you use the release linked to above given that it relies on importing the actual simulation code from the Gist.
+When I failed to find a simple solution, I created the [pyscript2gce](https://github.com/LukasWallrich/pyscript2gce-production) helper, which creates a Docker container that executes a script when launched and saves the results. Once set up, all it takes to run a script is to push an update to a specified file (here: `run_simulation.py`) and start up a VM with a single line terminal command. For that, [a GitHub action](https://github.com/LukasWallrich/diversity_abm_replication/blob/main/.github/workflows/push_gist.yml) pushes changes to that file automatically to a Gist, which is then accessed by the VM, based on [this version](https://github.com/LukasWallrich/pyscript2gce-production/releases/tag/Diversity-ABM-replication) of pyscript2gce. The README of pyscript2gce details how this can be set up. Note that you do not need to change anything in the Python code in `pyscript2gce` except for the link to the Gist in `run_simulation.py` if you use the release linked to above given that it relies on importing the actual simulation code from the Gist.
 
 
 
