### 2022-10-23 #61 @LukasWallrich
@oliviaguest I have now re-read the article, caught some typos and added the meta-data in as far as I have it (including the code DOI). The PDF is now [here](https://github.com/LukasWallrich/diversity_abm_replication-manuscript/raw/main/article.pdf) ... I will also update the link at the top. Please let me know if you catch anything else - and how I can get the article DOI and other remaining metadata (issue, volume etc).
