### 2022-12-19 #61 @rougier
@LukasWallrich I've uploaded a sandbox version at https://sandbox.zenodo.org/record/1137949. Can you check everythng's ok ? Note that **DOI is not final**, it'll change when I upload on the real Zenodo server.
