### 2022-12-19 #61 @LukasWallrich
Wonderful, thank you @rougier! Looks all good to me - the only thing is that citations and footnotes look exactly the same, though it is often clear from the context what the superscript refers to. If that is in line with your template, I'm ok with that ... if there is an easy fix, then that'd be good ...
