### 2022-12-20 #61 @LukasWallrich
I now looked at some recent articles and noticed that some use [1] etc rather than the superscript for references - which is the ReScience standard? If the square brackets, that'd resolve the footnote issue ... if so, would you happen to know how to change that in latex (np if not, Google will tell me eventually)
