### 2022-12-20 #61 @rougier
I never really paid attention to footnotes in the template and I think it should be the latex default. Did you use latest template?
