### 2022-12-20 #61 @LukasWallrich
Yes, I used the last template. I think the key question is about the references. Should they be [1] or <sup>1</sup>. Recent submissions use both formats, e.g., #67 uses superscripts and #69 uses square brackets. If they are <sup>1</sup>, then they clash with footnotes - e.g., in [10.5281/zenodo.6574651](http://dx.doi.org/10.5281/zenodo.6574651).

My suggestion would be to use a combination, namely <sup>[1]</sup>, for citations - in line with, e.g., Wikipedia - see [my new PDF here](https://github.com/LukasWallrich/diversity_abm_replication-manuscript/raw/main/article.pdf). To get that, I needed to make a small edit addition to [rescience.cls](https://github.com/LukasWallrich/diversity_abm_replication-manuscript/blob/main/rescience.cls#L325). Or what would be your preferred solution?
