### 2022-12-20 #61 @rougier
Seems to be a good compromise. Can you update the cls in your repo (I'll update my fork) and possibly make PR for the template ?
