### 2022-12-21 #61 @LukasWallrich
Yes - have submitted the PR and updated my cls - hope this will enable you to remake the PDF accordingly. Let me know if you come across any issues - thanks!
