+++ manuscript/metadata.tex
@@ -5,34 +5,34 @@
 \def \codeSWH{}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
-\def \reviewerIINAME{}
-\def \reviewerIIORCID{}
-\def \dateRECEIVED{01 November 2018}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \editorNAME{Olivia Guest}
+\def \editorORCID{0000-0002-1891-0972}
+\def \reviewerINAME{Marieke Woensdregt}
+\def \reviewerIORCID{0000-0001-5454-4722}
+\def \reviewerIINAME{Fausto Carcassi}
+\def \reviewerIIORCID{0000-0001-6843-1737}
+\def \dateRECEIVED{29 November 2021}
+\def \dateACCEPTED{19 October 2022}
+\def \datePUBLISHED{19 December 2022}
 \def \articleTITLE{[Re] Groups of diverse problem-solvers outperform groups of highest-ability problem-solvers - most of the time}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{Social Psychology}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
-\def \articleYEAR{2021}
-\def \reviewURL{}
-\def \articleABSTRACT{Problem-solving teams are becoming increasingly diverse, which has been suggested to improve performance. Pioneering computational work by Hong \& Page (2004) suggested that diversity indeed trumps ability when it comes to finding the peaks in a random landscape. Recently, Grim and colleagues (2019) extended the model to consider more structured landscapes and suggested that the original claim only holds under limited circumstances. In this paper, I replicate the main findings of the two works, and provide modular, extensible code to facilitate future research into this paradigm. Overall, the replications were successful, even though a closer look at the data suggests that the limitations highlighted by Grim and colleagues are less severe than initially presented, so that diversity on the whole still appears to trump ability within this paradigm. The code for the agent-based models and my analyses is available on https://github.com/LukasWallrich/diversity abm replication.}
+\def \articleYEAR{2022}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/61}
+\def \articleABSTRACT{Problem-solving teams are becoming increasingly diverse, which has been suggested to improve performance. Pioneering computational work by Hong \& Page (2004) suggested that diversity indeed trumps ability when it comes to finding the peaks in a random landscape. Recently, Grim and colleagues (2019) extended the model to consider more structured landscapes and suggested that the original claim only holds under limited circumstances. In this paper, I replicate the main findings of the two works, and provide modular, extensible code to facilitate future research into this paradigm. Overall, the replications were successful, even though a closer look at the data suggests that the limitations highlighted by Grim and colleagues are less severe than initially presented, so that diversity on the whole still appears to trump ability within this paradigm. The code for the agent-based models and my analyses is available on https://github.com/LukasWallrich/diversityÂ abmÂ replication.}
 \def \replicationCITE{Grim, P., Singer, D. J., Bramson, A., Holman, B., McGeehan, S., & Berger, W. J. (2019). Diversity, ability, and expertise in epistemic communities. Philosophy of Science, 86(1), 98-123.}
 \def \replicationBIB{grim_diversity_2019}
[…]
