### 2022-12-26 #61 @rougier
Here is the new (sandbox) entry: https://sandbox.zenodo.org/record/1139997. If everyhthing's fine, I'll publish it and make a PR on your repo.
