### 2022-12-26 #61 @LukasWallrich
Thanks - this looks good!

On Mon, 26 Dec 2022, 13:36 Nicolas P. Rougier, ***@***.***>
wrote:

> Here is the new (sandbox) entry: https://sandbox.zenodo.org/record/1139997.
> If everyhthing's fine, I'll publish it and make a PR on your repo.
>
