### 2022-12-26 #61 @rougier
Entry is online at https://zenodo.org/record/7484072 ! It will appear on rescience website in a few minutes
