+++ manuscript/metadata.tex
@@ -13,7 +13,7 @@
 \def \reviewerIIORCID{0000-0001-6843-1737}
 \def \dateRECEIVED{29 November 2021}
 \def \dateACCEPTED{19 October 2022}
-\def \datePUBLISHED{19 December 2022}
+\def \datePUBLISHED{26 December 2022}
 \def \articleTITLE{[Re] Groups of diverse problem-solvers outperform groups of highest-ability problem-solvers - most of the time}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{Social Psychology}
@@ -32,7 +32,7 @@
 \def \journalVOLUME{8}
 \def \journalISSUE{1}
 \def \articleNUMBER{6}
-\def \articleDOI{10.5072/zenodo.1137949}
+\def \articleDOI{10.5281/zenodo.7484072}
 \def \authorsFULL{Lukas Wallrich}
 \def \authorsABBRV{L. Wallrich}
 \def \authorsSHORT{Wallrich}
