+++ manuscript/metadata.yaml
@@ -94,13 +94,13 @@ contributors:
 dates:
   - received:  November 29, 2021
   - accepted:  October 19, 2022
-  - published: December 19, 2022
+  - published: December 26, 2022
 
 # This information will be provided by the editor
 article:
   - number: 6
-  - doi: 10.5072/zenodo.1137949
-  - url: https://sandbox.zenodo.org/record/1137949/files/article.pdf   
+  - doi: 10.5281/zenodo.7484072
+  - url: https://zenodo.org/record/7484072/files/article.pdf
 
 # This information will be provided by the editor
 journal:
