### 2022-12-26 #61 @rougier
Here: https://rescience.github.io/bibliography/Wallrich_2022.html
